+++ CHANGELOG.md
@@ -27,3 +27,4 @@
 * (ERICK) Created `ModuleHandler` class.
 * (ERICK) Created `refreshModuleHandler()` method.
 * (ERICK) Added lists for Take and Teach.
+* (ERICK) Created `addStudentToModule()` method.