+++ CHANGELOG.md
@@ -31,3 +31,4 @@
 * (ERICK) Created `addStaffToModule()` method.
 * (ERICK) Created `removeStudentFromModule()` method.
 * (ERICK) Created `removeStaffFromModule()` method.
+* (ERICK) Student-take not created if already exists.